refactor(executor): Split columnar.rs (1505 lines) into smaller modules
Split hash_join/columnar.rs into a directory with logical submodules:
- hash_table.rs (388 lines): Hash table structures for single/multi-column keys
- probe.rs (206 lines): Probe phase implementations for inner and outer joins
- output.rs (396 lines): Result construction and column gathering
- row_extract.rs (212 lines): Row-to-column extraction utilities
- mod.rs (335 lines): Main entry points and re-exports

All existing tests pass. No public API changes.

Closes #3054
