--- crates/vibesql-executor/src/select/join/hash_join/columnar.rs
@@ -1,1505 +1,0 @@
-//! Columnar Hash Join Implementation
-//!
-//! This module provides a high-performance hash join that operates entirely on
-//! columnar data, avoiding row materialization overhead.
-//!
-//! ## Performance Characteristics
-//!
-//! - **Build phase**: O(n) with SIMD hash computation on contiguous arrays
-//! - **Probe phase**: O(m) with SIMD hash computation and vectorized lookups
-//! - **Memory**: O(n) for hash table + O(result_size) for output
-//!
-//! ## Key Optimizations
-//!
-//! 1. **No SqlValue enum dispatch**: Operates directly on typed arrays (i64, f64, etc.)
-//! 2. **SIMD hashing**: Hashes 4-8 values simultaneously
-//! 3. **Cache-friendly**: Contiguous memory access patterns
-//! 4. **Deferred materialization**: Only creates rows at output if needed
-//!
-//! Note: This module is experimental/research code. Some functions are not yet
-//! integrated into the main execution path.
-
-#![allow(dead_code)]
-
-use crate::errors::ExecutorError;
-use crate::select::columnar::{ColumnArray, ColumnarBatch};
-
-/// Hash table for columnar join operations
-///
-/// Uses a two-level structure:
-/// - `buckets`: Maps hash value to first entry index
-/// - `entries`: Linked list of (row_index, next_entry) pairs
-///
-/// This structure is more cache-friendly than HashMap<Key, Vec<usize>> because:
-/// - No per-bucket Vec allocation
-/// - Entries are stored contiguously
-/// - Better cache utilization during probe
-pub struct ColumnarHashTable {
-    /// Number of hash buckets (power of 2)
-    bucket_count: usize,
-    /// Bucket array: bucket[hash % bucket_count] = first entry index (or u32::MAX if empty)
-    buckets: Vec<u32>,
-    /// Entry array: entries[i] = (row_index, next_entry_index)
-    entries: Vec<(u32, u32)>,
-}
-
-impl ColumnarHashTable {
-    /// Build a hash table from an integer column
-    ///
-    /// This is the fast path for integer-keyed joins (most common in TPC-H).
-    pub fn build_from_i64(values: &[i64]) -> Self {
-        let row_count = values.len();
-
-        // Size buckets to ~2x entries for good load factor
-        let bucket_count = (row_count * 2).next_power_of_two().max(16);
-        let bucket_mask = bucket_count - 1;
-
-        // Initialize buckets to empty (u32::MAX)
-        let mut buckets = vec![u32::MAX; bucket_count];
-
-        // Pre-allocate entries
-        let mut entries = Vec::with_capacity(row_count);
-
-        // Build hash table
-        for (row_idx, &value) in values.iter().enumerate() {
-            // Simple hash for i64: mix the bits
-            let hash = Self::hash_i64(value);
-            let bucket_idx = (hash as usize) & bucket_mask;
-
-            // Insert into linked list at this bucket
-            let prev_head = buckets[bucket_idx];
-            entries.push((row_idx as u32, prev_head));
-            buckets[bucket_idx] = entries.len() as u32 - 1;
-        }
-
-        Self { bucket_count, buckets, entries }
-    }
-
-    /// Build a hash table from a string column
-    pub fn build_from_string(values: &[String]) -> Self {
-        let row_count = values.len();
-        let bucket_count = (row_count * 2).next_power_of_two().max(16);
-        let bucket_mask = bucket_count - 1;
-
-        let mut buckets = vec![u32::MAX; bucket_count];
-        let mut entries = Vec::with_capacity(row_count);
-
-        for (row_idx, value) in values.iter().enumerate() {
-            let hash = Self::hash_string(value);
-            let bucket_idx = (hash as usize) & bucket_mask;
-
-            let prev_head = buckets[bucket_idx];
-            entries.push((row_idx as u32, prev_head));
-            buckets[bucket_idx] = entries.len() as u32 - 1;
-        }
-
-        Self { bucket_count, buckets, entries }
-    }
-
-    /// Build hash table from a ColumnArray
-    pub fn build_from_column(column: &ColumnArray) -> Result<Self, ExecutorError> {
-        match column {
-            ColumnArray::Int64(values, _nulls) => Ok(Self::build_from_i64(values)),
-            ColumnArray::Float64(values, _nulls) => Ok(Self::build_from_f64(values)),
-            ColumnArray::String(values, _nulls) => Ok(Self::build_from_string(values)),
-            ColumnArray::Date(values, _nulls) => Ok(Self::build_from_i32(values)),
-            ColumnArray::Timestamp(values, _nulls) => Ok(Self::build_from_i64(values)),
-            _ => Err(ExecutorError::UnsupportedFeature(
-                "Columnar hash join not supported for this column type".to_string()
-            )),
-        }
-    }
-
-    /// Build from i32 values (dates)
-    fn build_from_i32(values: &[i32]) -> Self {
-        let row_count = values.len();
-        let bucket_count = (row_count * 2).next_power_of_two().max(16);
-        let bucket_mask = bucket_count - 1;
-
-        let mut buckets = vec![u32::MAX; bucket_count];
-        let mut entries = Vec::with_capacity(row_count);
-
-        for (row_idx, &value) in values.iter().enumerate() {
-            let hash = Self::hash_i64(value as i64);
-            let bucket_idx = (hash as usize) & bucket_mask;
-
-            let prev_head = buckets[bucket_idx];
-            entries.push((row_idx as u32, prev_head));
-            buckets[bucket_idx] = entries.len() as u32 - 1;
-        }
-
-        Self { bucket_count, buckets, entries }
-    }
-
-    /// Build from f64 values
-    fn build_from_f64(values: &[f64]) -> Self {
-        let row_count = values.len();
-        let bucket_count = (row_count * 2).next_power_of_two().max(16);
-        let bucket_mask = bucket_count - 1;
-
-        let mut buckets = vec![u32::MAX; bucket_count];
-        let mut entries = Vec::with_capacity(row_count);
-
-        for (row_idx, &value) in values.iter().enumerate() {
-            let hash = Self::hash_f64(value);
-            let bucket_idx = (hash as usize) & bucket_mask;
-
-            let prev_head = buckets[bucket_idx];
-            entries.push((row_idx as u32, prev_head));
-            buckets[bucket_idx] = entries.len() as u32 - 1;
-        }
-
-        Self { bucket_count, buckets, entries }
-    }
-
-    /// Probe the hash table with an i64 key, returning matching row indices
-    #[inline]
-    pub fn probe_i64<'a>(&'a self, key: i64, build_values: &'a [i64]) -> impl Iterator<Item = u32> + 'a {
-        let hash = Self::hash_i64(key);
-        let bucket_idx = (hash as usize) & (self.bucket_count - 1);
-
-        HashTableIter {
-            entries: &self.entries,
-            current: self.buckets[bucket_idx],
-            key_checker: move |row_idx: u32| build_values[row_idx as usize] == key,
-        }
-    }
-
-    /// Probe the hash table with a string key
-    #[inline]
-    pub fn probe_string<'a>(&'a self, key: &'a str, build_values: &'a [String]) -> impl Iterator<Item = u32> + 'a {
-        let hash = Self::hash_string(key);
-        let bucket_idx = (hash as usize) & (self.bucket_count - 1);
-
-        HashTableIter {
-            entries: &self.entries,
-            current: self.buckets[bucket_idx],
-            key_checker: move |row_idx: u32| build_values[row_idx as usize] == key,
-        }
-    }
-
-    /// Fast hash function for i64 (FxHash-style)
-    #[inline(always)]
-    fn hash_i64(value: i64) -> u64 {
-        const K: u64 = 0x517cc1b727220a95;
-        let mut h = value as u64;
-        h = h.wrapping_mul(K);
-        h ^= h >> 32;
-        h
-    }
-
-    /// Hash function for f64
-    #[inline(always)]
-    fn hash_f64(value: f64) -> u64 {
-        Self::hash_i64(value.to_bits() as i64)
-    }
-
-    /// Hash function for strings (FNV-1a style)
-    #[inline(always)]
-    fn hash_string(value: &str) -> u64 {
-        const FNV_OFFSET: u64 = 0xcbf29ce484222325;
-        const FNV_PRIME: u64 = 0x100000001b3;
-
-        let mut hash = FNV_OFFSET;
-        for byte in value.bytes() {
-            hash ^= byte as u64;
-            hash = hash.wrapping_mul(FNV_PRIME);
-        }
-        hash
-    }
-}
-
-/// Iterator over hash table entries matching a key
-struct HashTableIter<'a, F> {
-    entries: &'a [(u32, u32)],
-    current: u32,
-    key_checker: F,
-}
-
-impl<'a, F: Fn(u32) -> bool> Iterator for HashTableIter<'a, F> {
-    type Item = u32;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        while self.current != u32::MAX {
-            let (row_idx, next) = self.entries[self.current as usize];
-            self.current = next;
-
-            if (self.key_checker)(row_idx) {
-                return Some(row_idx);
-            }
-        }
-        None
-    }
-}
-
-/// Result of a columnar hash join probe phase
-pub struct JoinIndices {
-    /// Indices into left (probe) batch
-    pub left_indices: Vec<u32>,
-    /// Indices into right (build) batch
-    pub right_indices: Vec<u32>,
-}
-
-/// Execute a columnar inner hash join
-///
-/// This function operates entirely on columnar data without materializing rows.
-///
-/// # Arguments
-/// * `left_batch` - Left (probe) side columnar batch
-/// * `right_batch` - Right (build) side columnar batch
-/// * `left_key_idx` - Column index of join key in left batch
-/// * `right_key_idx` - Column index of join key in right batch
-///
-/// # Returns
-/// A new ColumnarBatch containing joined columns
-pub fn columnar_hash_join_inner(
-    left_batch: &ColumnarBatch,
-    right_batch: &ColumnarBatch,
-    left_key_idx: usize,
-    right_key_idx: usize,
-) -> Result<ColumnarBatch, ExecutorError> {
-    // Get key columns
-    let left_key = left_batch.column(left_key_idx)
-        .ok_or_else(|| ExecutorError::ColumnarColumnNotFound {
-            column_index: left_key_idx,
-            batch_columns: left_batch.column_count(),
-        })?;
-    let right_key = right_batch.column(right_key_idx)
-        .ok_or_else(|| ExecutorError::ColumnarColumnNotFound {
-            column_index: right_key_idx,
-            batch_columns: right_batch.column_count(),
-        })?;
-
-    // Build hash table on right (smaller side ideally)
-    // For now, always build on right - could optimize to choose smaller side
-    let hash_table = ColumnarHashTable::build_from_column(right_key)?;
-
-    // Probe and collect matching indices
-    let join_indices = probe_columnar(&hash_table, left_key, right_key)?;
-
-    // Gather result columns
-    gather_join_result(left_batch, right_batch, &join_indices)
-}
-
-/// Check if a value is NULL according to the null bitmap
-#[inline]
-fn is_null(nulls: &Option<std::sync::Arc<Vec<bool>>>, idx: usize) -> bool {
-    nulls.as_ref().is_some_and(|n| n.get(idx).copied().unwrap_or(false))
-}
-
-/// Probe phase: find all matching pairs
-///
-/// NULL handling: NULL keys never match in equi-joins (NULL = NULL is NULL, not true).
-/// Both left and right NULL keys are skipped during probe.
-fn probe_columnar(
-    hash_table: &ColumnarHashTable,
-    left_key: &ColumnArray,
-    right_key: &ColumnArray,
-) -> Result<JoinIndices, ExecutorError> {
-    let mut left_indices = Vec::new();
-    let mut right_indices = Vec::new();
-
-    match (left_key, right_key) {
-        (ColumnArray::Int64(left_values, left_nulls), ColumnArray::Int64(right_values, right_nulls)) => {
-            for (left_idx, &key) in left_values.iter().enumerate() {
-                // Skip NULL left keys - NULLs never match in equi-joins
-                if is_null(left_nulls, left_idx) {
-                    continue;
-                }
-                for right_idx in hash_table.probe_i64(key, right_values) {
-                    // Skip NULL right keys
-                    if is_null(right_nulls, right_idx as usize) {
-                        continue;
-                    }
-                    left_indices.push(left_idx as u32);
-                    right_indices.push(right_idx);
-                }
-            }
-        }
-        (ColumnArray::String(left_values, left_nulls), ColumnArray::String(right_values, right_nulls)) => {
-            for (left_idx, key) in left_values.iter().enumerate() {
-                // Skip NULL left keys - NULLs never match in equi-joins
-                if is_null(left_nulls, left_idx) {
-                    continue;
-                }
-                for right_idx in hash_table.probe_string(key, right_values) {
-                    // Skip NULL right keys
-                    if is_null(right_nulls, right_idx as usize) {
-                        continue;
-                    }
-                    left_indices.push(left_idx as u32);
-                    right_indices.push(right_idx);
-                }
-            }
-        }
-        _ => {
-            return Err(ExecutorError::UnsupportedFeature(
-                "Columnar hash join probe not supported for this column type combination".to_string()
-            ));
-        }
-    }
-
-    Ok(JoinIndices { left_indices, right_indices })
-}
-
-/// Gather columns from both batches based on join indices
-fn gather_join_result(
-    left_batch: &ColumnarBatch,
-    right_batch: &ColumnarBatch,
-    indices: &JoinIndices,
-) -> Result<ColumnarBatch, ExecutorError> {
-    let _result_count = indices.left_indices.len();
-
-    // Gather left columns
-    let mut result_columns = Vec::new();
-    for col_idx in 0..left_batch.column_count() {
-        let column = left_batch.column(col_idx).unwrap();
-        let gathered = gather_column(column, &indices.left_indices)?;
-        result_columns.push(gathered);
-    }
-
-    // Gather right columns
-    for col_idx in 0..right_batch.column_count() {
-        let column = right_batch.column(col_idx).unwrap();
-        let gathered = gather_column(column, &indices.right_indices)?;
-        result_columns.push(gathered);
-    }
-
-    // Combine column names
-    let column_names = match (left_batch.column_names(), right_batch.column_names()) {
-        (Some(left_names), Some(right_names)) => {
-            let mut names = left_names.to_vec();
-            names.extend(right_names.iter().cloned());
-            Some(names)
-        }
-        _ => None,
-    };
-
-    ColumnarBatch::from_columns(result_columns, column_names)
-}
-
-/// Gather values from a column based on indices
-fn gather_column(column: &ColumnArray, indices: &[u32]) -> Result<ColumnArray, ExecutorError> {
-    match column {
-        ColumnArray::Int64(values, nulls) => {
-            let gathered: Vec<i64> = indices.iter()
-                .map(|&idx| values[idx as usize])
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::Int64(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::Int32(values, nulls) => {
-            let gathered: Vec<i32> = indices.iter()
-                .map(|&idx| values[idx as usize])
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::Int32(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::Float64(values, nulls) => {
-            let gathered: Vec<f64> = indices.iter()
-                .map(|&idx| values[idx as usize])
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::Float64(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::Float32(values, nulls) => {
-            let gathered: Vec<f32> = indices.iter()
-                .map(|&idx| values[idx as usize])
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::Float32(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::String(values, nulls) => {
-            let gathered: Vec<String> = indices.iter()
-                .map(|&idx| values[idx as usize].clone())
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::String(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::FixedString(values, nulls) => {
-            let gathered: Vec<String> = indices.iter()
-                .map(|&idx| values[idx as usize].clone())
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::FixedString(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::Date(values, nulls) => {
-            let gathered: Vec<i32> = indices.iter()
-                .map(|&idx| values[idx as usize])
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::Date(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::Timestamp(values, nulls) => {
-            let gathered: Vec<i64> = indices.iter()
-                .map(|&idx| values[idx as usize])
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::Timestamp(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::Boolean(values, nulls) => {
-            let gathered: Vec<u8> = indices.iter()
-                .map(|&idx| values[idx as usize])
-                .collect();
-            let gathered_nulls = nulls.as_ref().map(|n| {
-                std::sync::Arc::new(indices.iter().map(|&idx| n[idx as usize]).collect())
-            });
-            Ok(ColumnArray::Boolean(std::sync::Arc::new(gathered), gathered_nulls))
-        }
-        ColumnArray::Mixed(values) => {
-            let gathered: Vec<vibesql_types::SqlValue> = indices.iter()
-                .map(|&idx| values[idx as usize].clone())
-                .collect();
-            Ok(ColumnArray::Mixed(std::sync::Arc::new(gathered)))
-        }
-    }
-}
-
-/// Extract a single column from rows as a typed array (for integer columns)
-///
-/// This enables using columnar hash operations on row-based data.
-/// Returns None if the column contains non-integer values or NULLs.
-pub fn extract_i64_column(rows: &[vibesql_storage::Row], col_idx: usize) -> Option<Vec<i64>> {
-    let mut values = Vec::with_capacity(rows.len());
-
-    for row in rows {
-        match row.values.get(col_idx) {
-            Some(vibesql_types::SqlValue::Integer(v)) => values.push(*v),
-            Some(vibesql_types::SqlValue::Bigint(v)) => values.push(*v),
-            Some(vibesql_types::SqlValue::Smallint(v)) => values.push(*v as i64),
-            _ => return None, // Non-integer or NULL value
-        }
-    }
-
-    Some(values)
-}
-
-/// Hash join using columnar hash table on row-based data
-///
-/// This function provides a fast path for integer equi-joins by:
-/// 1. Extracting join columns as typed i64 arrays
-/// 2. Using the columnar hash table for O(1) lookups without SqlValue dispatch
-/// 3. Returning index pairs for row combination
-///
-/// Returns None if the join columns are not integer types.
-pub fn hash_join_indices_columnar(
-    build_rows: &[vibesql_storage::Row],
-    probe_rows: &[vibesql_storage::Row],
-    build_col_idx: usize,
-    probe_col_idx: usize,
-) -> Option<Vec<(usize, usize)>> {
-    // Extract join columns as typed arrays
-    let build_keys = extract_i64_column(build_rows, build_col_idx)?;
-    let probe_keys = extract_i64_column(probe_rows, probe_col_idx)?;
-
-    // Build hash table on build side
-    let hash_table = ColumnarHashTable::build_from_i64(&build_keys);
-
-    // Probe and collect matching index pairs
-    let estimated_capacity = probe_keys.len().min(100_000);
-    let mut join_pairs = Vec::with_capacity(estimated_capacity);
-
-    for (probe_idx, &probe_key) in probe_keys.iter().enumerate() {
-        for build_idx in hash_table.probe_i64(probe_key, &build_keys) {
-            join_pairs.push((build_idx as usize, probe_idx));
-        }
-    }
-
-    Some(join_pairs)
-}
-
-/// Hash table for multi-column integer joins
-///
-/// Uses pre-computed composite hashes for efficient multi-column lookups.
-/// This avoids SqlValue enum dispatch and Vec allocation for each key.
-pub struct CompositeIntHashTable {
-    /// Number of hash buckets (power of 2)
-    bucket_count: usize,
-    /// Bucket array: bucket[hash % bucket_count] = first entry index (or u32::MAX if empty)
-    buckets: Vec<u32>,
-    /// Entry array: entries[i] = (row_index, next_entry_index)
-    entries: Vec<(u32, u32)>,
-    /// Number of key columns
-    num_cols: usize,
-}
-
-impl CompositeIntHashTable {
-    /// Hash multiple i64 values into a single u64
-    #[inline(always)]
-    fn hash_composite(values: &[i64]) -> u64 {
-        const K: u64 = 0x517cc1b727220a95;
-        let mut h: u64 = 0xcbf29ce484222325; // FNV offset basis
-
-        for &v in values {
-            h ^= v as u64;
-            h = h.wrapping_mul(K);
-            h ^= h >> 32;
-        }
-        h
-    }
-
-    /// Build a hash table from multiple integer columns
-    ///
-    /// Fast path for multi-column integer-keyed joins (common in TPC-H Q3, Q7, Q10).
-    pub fn build_from_multi_i64(columns: &[&[i64]]) -> Self {
-        if columns.is_empty() {
-            return Self {
-                bucket_count: 16,
-                buckets: vec![u32::MAX; 16],
-                entries: Vec::new(),
-                num_cols: 0,
-            };
-        }
-
-        let row_count = columns[0].len();
-        let num_cols = columns.len();
-
-        // Size buckets to ~2x entries for good load factor
-        let bucket_count = (row_count * 2).next_power_of_two().max(16);
-        let bucket_mask = bucket_count - 1;
-
-        // Initialize buckets to empty (u32::MAX)
-        let mut buckets = vec![u32::MAX; bucket_count];
-
-        // Pre-allocate entries
-        let mut entries = Vec::with_capacity(row_count);
-
-        // Build hash table
-        let mut key_values = vec![0i64; num_cols];
-        for row_idx in 0..row_count {
-            // Extract key values for this row
-            for (col_idx, col) in columns.iter().enumerate() {
-                key_values[col_idx] = col[row_idx];
-            }
-
-            let hash = Self::hash_composite(&key_values);
-            let bucket_idx = (hash as usize) & bucket_mask;
-
-            // Insert into linked list at this bucket
-            let prev_head = buckets[bucket_idx];
-            entries.push((row_idx as u32, prev_head));
-            buckets[bucket_idx] = entries.len() as u32 - 1;
-        }
-
-        Self { bucket_count, buckets, entries, num_cols }
-    }
-
-    /// Probe the hash table with multiple i64 keys
-    #[inline]
-    pub fn probe_multi_i64<'a>(
-        &'a self,
-        probe_keys: &'a [i64],
-        build_columns: &'a [&'a [i64]],
-    ) -> impl Iterator<Item = u32> + 'a {
-        let hash = Self::hash_composite(probe_keys);
-        let bucket_idx = (hash as usize) & (self.bucket_count - 1);
-        let num_cols = self.num_cols;
-
-        CompositeHashTableIter {
-            entries: &self.entries,
-            current: self.buckets[bucket_idx],
-            probe_keys,
-            build_columns,
-            num_cols,
-        }
-    }
-}
-
-/// Iterator over composite hash table entries matching a multi-column key
-struct CompositeHashTableIter<'a> {
-    entries: &'a [(u32, u32)],
-    current: u32,
-    probe_keys: &'a [i64],
-    build_columns: &'a [&'a [i64]],
-    num_cols: usize,
-}
-
-impl<'a> Iterator for CompositeHashTableIter<'a> {
-    type Item = u32;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        while self.current != u32::MAX {
-            let (row_idx, next) = self.entries[self.current as usize];
-            self.current = next;
-
-            // Check if all columns match
-            let mut matches = true;
-            for col_idx in 0..self.num_cols {
-                if self.build_columns[col_idx][row_idx as usize] != self.probe_keys[col_idx] {
-                    matches = false;
-                    break;
-                }
-            }
-
-            if matches {
-                return Some(row_idx);
-            }
-        }
-        None
-    }
-}
-
-/// Extract multiple integer columns from rows as typed arrays
-///
-/// Returns None if any column contains non-integer values or NULLs.
-pub fn extract_multi_i64_columns(
-    rows: &[vibesql_storage::Row],
-    col_indices: &[usize],
-) -> Option<Vec<Vec<i64>>> {
-    let mut columns: Vec<Vec<i64>> = col_indices.iter().map(|_| Vec::with_capacity(rows.len())).collect();
-
-    for row in rows {
-        for (out_idx, &col_idx) in col_indices.iter().enumerate() {
-            match row.values.get(col_idx) {
-                Some(vibesql_types::SqlValue::Integer(v)) => columns[out_idx].push(*v),
-                Some(vibesql_types::SqlValue::Bigint(v)) => columns[out_idx].push(*v),
-                Some(vibesql_types::SqlValue::Smallint(v)) => columns[out_idx].push(*v as i64),
-                _ => return None, // Non-integer or NULL value
-            }
-        }
-    }
-
-    Some(columns)
-}
-
-/// Multi-column hash join using columnar hash table on row-based data
-///
-/// This function provides a fast path for multi-column integer equi-joins by:
-/// 1. Extracting join columns as typed i64 arrays (avoiding SqlValue enum dispatch)
-/// 2. Using the composite columnar hash table with pre-computed hashes
-/// 3. Returning index pairs for row combination
-///
-/// Returns None if any join columns are not integer types.
-pub fn hash_join_indices_columnar_multi(
-    build_rows: &[vibesql_storage::Row],
-    probe_rows: &[vibesql_storage::Row],
-    build_col_indices: &[usize],
-    probe_col_indices: &[usize],
-) -> Option<Vec<(usize, usize)>> {
-    // Extract join columns as typed arrays
-    let build_columns = extract_multi_i64_columns(build_rows, build_col_indices)?;
-    let probe_columns = extract_multi_i64_columns(probe_rows, probe_col_indices)?;
-
-    // Convert to slice references for the hash table
-    let build_col_refs: Vec<&[i64]> = build_columns.iter().map(|c| c.as_slice()).collect();
-    let probe_col_refs: Vec<&[i64]> = probe_columns.iter().map(|c| c.as_slice()).collect();
-
-    // Build hash table on build side
-    let hash_table = CompositeIntHashTable::build_from_multi_i64(&build_col_refs);
-
-    // Probe and collect matching index pairs
-    let estimated_capacity = probe_rows.len().min(100_000);
-    let mut join_pairs = Vec::with_capacity(estimated_capacity);
-
-    let num_cols = probe_col_indices.len();
-    let mut probe_key = vec![0i64; num_cols];
-
-    for probe_idx in 0..probe_rows.len() {
-        // Extract probe keys for this row
-        for (col_idx, col) in probe_col_refs.iter().enumerate() {
-            probe_key[col_idx] = col[probe_idx];
-        }
-
-        for build_idx in hash_table.probe_multi_i64(&probe_key, &build_col_refs) {
-            join_pairs.push((build_idx as usize, probe_idx));
-        }
-    }
-
-    Some(join_pairs)
-}
-
-/// Execute a columnar LEFT OUTER hash join
-///
-/// This function operates entirely on columnar data without materializing rows.
-/// LEFT OUTER JOIN preserves all rows from the left (probe) side, outputting
-/// NULL values for right columns when there's no match.
-///
-/// # Arguments
-/// * `left_batch` - Left (probe) side columnar batch - all rows preserved
-/// * `right_batch` - Right (build) side columnar batch
-/// * `left_key_idx` - Column index of join key in left batch
-/// * `right_key_idx` - Column index of join key in right batch
-///
-/// # Returns
-/// A new ColumnarBatch containing joined columns with left rows preserved
-pub fn columnar_hash_join_left_outer(
-    left_batch: &ColumnarBatch,
-    right_batch: &ColumnarBatch,
-    left_key_idx: usize,
-    right_key_idx: usize,
-) -> Result<ColumnarBatch, ExecutorError> {
-    // Get key columns
-    let left_key = left_batch.column(left_key_idx)
-        .ok_or_else(|| ExecutorError::ColumnarColumnNotFound {
-            column_index: left_key_idx,
-            batch_columns: left_batch.column_count(),
-        })?;
-    let right_key = right_batch.column(right_key_idx)
-        .ok_or_else(|| ExecutorError::ColumnarColumnNotFound {
-            column_index: right_key_idx,
-            batch_columns: right_batch.column_count(),
-        })?;
-
-    // Build hash table on right side
-    let hash_table = ColumnarHashTable::build_from_column(right_key)?;
-
-    // Probe and collect matching indices, tracking unmatched left rows
-    let join_indices = probe_columnar_left_outer(&hash_table, left_key, right_key, left_batch.row_count())?;
-
-    // Gather result columns with NULL handling for unmatched rows
-    gather_left_outer_result(left_batch, right_batch, &join_indices)
-}
-
-/// Execute a columnar RIGHT OUTER hash join
-///
-/// This function operates entirely on columnar data without materializing rows.
-/// RIGHT OUTER JOIN preserves all rows from the right (build) side, outputting
-/// NULL values for left columns when there's no match.
-///
-/// # Arguments
-/// * `left_batch` - Left (probe) side columnar batch
-/// * `right_batch` - Right (build) side columnar batch - all rows preserved
-/// * `left_key_idx` - Column index of join key in left batch
-/// * `right_key_idx` - Column index of join key in right batch
-///
-/// # Returns
-/// A new ColumnarBatch containing joined columns with right rows preserved
-pub fn columnar_hash_join_right_outer(
-    left_batch: &ColumnarBatch,
-    right_batch: &ColumnarBatch,
-    left_key_idx: usize,
-    right_key_idx: usize,
-) -> Result<ColumnarBatch, ExecutorError> {
-    // Get key columns
-    let left_key = left_batch.column(left_key_idx)
-        .ok_or_else(|| ExecutorError::ColumnarColumnNotFound {
-            column_index: left_key_idx,
-            batch_columns: left_batch.column_count(),
-        })?;
-    let right_key = right_batch.column(right_key_idx)
-        .ok_or_else(|| ExecutorError::ColumnarColumnNotFound {
-            column_index: right_key_idx,
-            batch_columns: right_batch.column_count(),
-        })?;
-
-    // Build hash table on left side (reverse of inner join)
-    let hash_table = ColumnarHashTable::build_from_column(left_key)?;
-
-    // Probe with right side and track unmatched right rows
-    let join_indices = probe_columnar_right_outer(&hash_table, right_key, left_key, right_batch.row_count())?;
-
-    // Gather result columns with NULL handling for unmatched rows
-    gather_right_outer_result(left_batch, right_batch, &join_indices)
-}
-
-/// Result indices for LEFT OUTER join
-pub struct LeftOuterJoinIndices {
-    /// Indices into left (probe) batch - always valid
-    pub left_indices: Vec<u32>,
-    /// Indices into right (build) batch - u32::MAX means no match (NULL row)
-    pub right_indices: Vec<u32>,
-}
-
-/// Result indices for RIGHT OUTER join
-pub struct RightOuterJoinIndices {
-    /// Indices into left (probe) batch - u32::MAX means no match (NULL row)
-    pub left_indices: Vec<u32>,
-    /// Indices into right (build) batch - always valid
-    pub right_indices: Vec<u32>,
-}
-
-/// Probe phase for LEFT OUTER join: find matches and preserve unmatched left rows
-fn probe_columnar_left_outer(
-    hash_table: &ColumnarHashTable,
-    left_key: &ColumnArray,
-    right_key: &ColumnArray,
-    left_row_count: usize,
-) -> Result<LeftOuterJoinIndices, ExecutorError> {
-    let mut left_indices = Vec::new();
-    let mut right_indices = Vec::new();
-    let mut left_matched = vec![false; left_row_count];
-
-    match (left_key, right_key) {
-        (ColumnArray::Int64(left_values, left_nulls), ColumnArray::Int64(right_values, _)) => {
-            for (left_idx, &key) in left_values.iter().enumerate() {
-                // Skip NULL keys - they never match but still output with NULLs
-                let is_null = left_nulls.as_ref().map(|n| n[left_idx]).unwrap_or(false);
-                if is_null {
-                    continue; // Will be handled as unmatched
-                }
-
-                let mut found_match = false;
-                for right_idx in hash_table.probe_i64(key, right_values) {
-                    left_indices.push(left_idx as u32);
-                    right_indices.push(right_idx);
-                    found_match = true;
-                }
-                if found_match {
-                    left_matched[left_idx] = true;
-                }
-            }
-        }
-        (ColumnArray::String(left_values, left_nulls), ColumnArray::String(right_values, _)) => {
-            for (left_idx, key) in left_values.iter().enumerate() {
-                let is_null = left_nulls.as_ref().map(|n| n[left_idx]).unwrap_or(false);
-                if is_null {
-                    continue;
-                }
-
-                let mut found_match = false;
-                for right_idx in hash_table.probe_string(key, right_values) {
-                    left_indices.push(left_idx as u32);
-                    right_indices.push(right_idx);
-                    found_match = true;
-                }
-                if found_match {
-                    left_matched[left_idx] = true;
-                }
-            }
-        }
-        _ => {
-            return Err(ExecutorError::UnsupportedFeature(
-                "Columnar LEFT OUTER hash join probe not supported for this column type combination".to_string()
-            ));
-        }
-    }
-
-    // Add unmatched left rows with NULL marker for right side
-    for (left_idx, &matched) in left_matched.iter().enumerate() {
-        if !matched {
-            left_indices.push(left_idx as u32);
-            right_indices.push(u32::MAX); // NULL marker
-        }
-    }
-
-    Ok(LeftOuterJoinIndices { left_indices, right_indices })
-}
-
-/// Probe phase for RIGHT OUTER join: find matches and preserve unmatched right rows
-fn probe_columnar_right_outer(
-    hash_table: &ColumnarHashTable,
-    right_key: &ColumnArray,
-    left_key: &ColumnArray,
-    right_row_count: usize,
-) -> Result<RightOuterJoinIndices, ExecutorError> {
-    let mut left_indices = Vec::new();
-    let mut right_indices = Vec::new();
-    let mut right_matched = vec![false; right_row_count];
-
-    match (right_key, left_key) {
-        (ColumnArray::Int64(right_values, right_nulls), ColumnArray::Int64(left_values, _)) => {
-            for (right_idx, &key) in right_values.iter().enumerate() {
-                let is_null = right_nulls.as_ref().map(|n| n[right_idx]).unwrap_or(false);
-                if is_null {
-                    continue;
-                }
-
-                let mut found_match = false;
-                for left_idx in hash_table.probe_i64(key, left_values) {
-                    left_indices.push(left_idx);
-                    right_indices.push(right_idx as u32);
-                    found_match = true;
-                }
-                if found_match {
-                    right_matched[right_idx] = true;
-                }
-            }
-        }
-        (ColumnArray::String(right_values, right_nulls), ColumnArray::String(left_values, _)) => {
-            for (right_idx, key) in right_values.iter().enumerate() {
-                let is_null = right_nulls.as_ref().map(|n| n[right_idx]).unwrap_or(false);
-                if is_null {
-                    continue;
-                }
-
-                let mut found_match = false;
-                for left_idx in hash_table.probe_string(key, left_values) {
-                    left_indices.push(left_idx);
-                    right_indices.push(right_idx as u32);
-                    found_match = true;
-                }
-                if found_match {
-                    right_matched[right_idx] = true;
-                }
-            }
-        }
-        _ => {
-            return Err(ExecutorError::UnsupportedFeature(
-                "Columnar RIGHT OUTER hash join probe not supported for this column type combination".to_string()
-            ));
-        }
-    }
-
-    // Add unmatched right rows with NULL marker for left side
-    for (right_idx, &matched) in right_matched.iter().enumerate() {
-        if !matched {
-            left_indices.push(u32::MAX); // NULL marker
-            right_indices.push(right_idx as u32);
-        }
-    }
-
-    Ok(RightOuterJoinIndices { left_indices, right_indices })
-}
-
-/// Gather result columns for LEFT OUTER join with NULL handling
-fn gather_left_outer_result(
-    left_batch: &ColumnarBatch,
-    right_batch: &ColumnarBatch,
-    indices: &LeftOuterJoinIndices,
-) -> Result<ColumnarBatch, ExecutorError> {
-    let mut result_columns = Vec::new();
-
-    // Gather left columns (all indices are valid)
-    for col_idx in 0..left_batch.column_count() {
-        let column = left_batch.column(col_idx).unwrap();
-        let gathered = gather_column(column, &indices.left_indices)?;
-        result_columns.push(gathered);
-    }
-
-    // Gather right columns with NULL handling for unmatched rows
-    for col_idx in 0..right_batch.column_count() {
-        let column = right_batch.column(col_idx).unwrap();
-        let gathered = gather_column_with_nulls(column, &indices.right_indices)?;
-        result_columns.push(gathered);
-    }
-
-    // Combine column names
-    let column_names = match (left_batch.column_names(), right_batch.column_names()) {
-        (Some(left_names), Some(right_names)) => {
-            let mut names = left_names.to_vec();
-            names.extend(right_names.iter().cloned());
-            Some(names)
-        }
-        _ => None,
-    };
-
-    ColumnarBatch::from_columns(result_columns, column_names)
-}
-
-/// Gather result columns for RIGHT OUTER join with NULL handling
-fn gather_right_outer_result(
-    left_batch: &ColumnarBatch,
-    right_batch: &ColumnarBatch,
-    indices: &RightOuterJoinIndices,
-) -> Result<ColumnarBatch, ExecutorError> {
-    let mut result_columns = Vec::new();
-
-    // Gather left columns with NULL handling for unmatched rows
-    for col_idx in 0..left_batch.column_count() {
-        let column = left_batch.column(col_idx).unwrap();
-        let gathered = gather_column_with_nulls(column, &indices.left_indices)?;
-        result_columns.push(gathered);
-    }
-
-    // Gather right columns (all indices are valid)
-    for col_idx in 0..right_batch.column_count() {
-        let column = right_batch.column(col_idx).unwrap();
-        let gathered = gather_column(column, &indices.right_indices)?;
-        result_columns.push(gathered);
-    }
-
-    // Combine column names
-    let column_names = match (left_batch.column_names(), right_batch.column_names()) {
-        (Some(left_names), Some(right_names)) => {
-            let mut names = left_names.to_vec();
-            names.extend(right_names.iter().cloned());
-            Some(names)
-        }
-        _ => None,
-    };
-
-    ColumnarBatch::from_columns(result_columns, column_names)
-}
-
-/// Gather values from a column with NULL handling for outer joins
-///
-/// u32::MAX indices are converted to NULL values
-fn gather_column_with_nulls(column: &ColumnArray, indices: &[u32]) -> Result<ColumnArray, ExecutorError> {
-    match column {
-        ColumnArray::Int64(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(0); // placeholder
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize]);
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::Int64(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::Int32(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(0);
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize]);
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::Int32(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::Float64(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(0.0);
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize]);
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::Float64(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::Float32(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(0.0);
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize]);
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::Float32(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::String(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(String::new());
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize].clone());
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::String(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::FixedString(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(String::new());
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize].clone());
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::FixedString(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::Date(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(0);
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize]);
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::Date(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::Timestamp(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(0);
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize]);
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::Timestamp(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::Boolean(values, _existing_nulls) => {
-            let mut gathered = Vec::with_capacity(indices.len());
-            let mut nulls = Vec::with_capacity(indices.len());
-
-            for &idx in indices {
-                if idx == u32::MAX {
-                    gathered.push(0);
-                    nulls.push(true);
-                } else {
-                    gathered.push(values[idx as usize]);
-                    nulls.push(false);
-                }
-            }
-
-            Ok(ColumnArray::Boolean(std::sync::Arc::new(gathered), Some(std::sync::Arc::new(nulls))))
-        }
-        ColumnArray::Mixed(values) => {
-            let gathered: Vec<vibesql_types::SqlValue> = indices.iter()
-                .map(|&idx| {
-                    if idx == u32::MAX {
-                        vibesql_types::SqlValue::Null
-                    } else {
-                        values[idx as usize].clone()
-                    }
-                })
-                .collect();
-            Ok(ColumnArray::Mixed(std::sync::Arc::new(gathered)))
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::Arc;
-
-    #[test]
-    fn test_hash_table_build_and_probe() {
-        let values = vec![10i64, 20, 30, 20, 40, 20];
-        let ht = ColumnarHashTable::build_from_i64(&values);
-
-        // Probe for 20 - should find indices 1, 3, 5
-        let matches: Vec<u32> = ht.probe_i64(20, &values).collect();
-        assert_eq!(matches.len(), 3);
-        assert!(matches.contains(&1));
-        assert!(matches.contains(&3));
-        assert!(matches.contains(&5));
-
-        // Probe for 10 - should find index 0
-        let matches: Vec<u32> = ht.probe_i64(10, &values).collect();
-        assert_eq!(matches, vec![0]);
-
-        // Probe for 99 - should find nothing
-        let matches: Vec<u32> = ht.probe_i64(99, &values).collect();
-        assert!(matches.is_empty());
-    }
-
-    #[test]
-    fn test_columnar_hash_join() {
-        use std::sync::Arc;
-
-        // Create left batch: customer_id, name
-        let left_columns = vec![
-            ColumnArray::Int64(Arc::new(vec![1, 2, 3, 4]), None),
-            ColumnArray::String(Arc::new(vec!["Alice".into(), "Bob".into(), "Carol".into(), "Dave".into()]), None),
-        ];
-        let left_batch = ColumnarBatch::from_columns(left_columns, Some(vec!["customer_id".into(), "name".into()])).unwrap();
-
-        // Create right batch: order_id, customer_id, amount
-        let right_columns = vec![
-            ColumnArray::Int64(Arc::new(vec![101, 102, 103, 104, 105]), None),
-            ColumnArray::Int64(Arc::new(vec![1, 2, 1, 3, 2]), None),
-            ColumnArray::Float64(Arc::new(vec![100.0, 200.0, 150.0, 300.0, 250.0]), None),
-        ];
-        let right_batch = ColumnarBatch::from_columns(right_columns, Some(vec!["order_id".into(), "customer_id".into(), "amount".into()])).unwrap();
-
-        // Join on customer_id (left col 0 = right col 1)
-        let result = columnar_hash_join_inner(&left_batch, &right_batch, 0, 1).unwrap();
-
-        // Should have 5 result rows (Alice has 2 orders, Bob has 2, Carol has 1)
-        assert_eq!(result.row_count(), 5);
-
-        // Should have 5 columns (2 from left + 3 from right)
-        assert_eq!(result.column_count(), 5);
-    }
-
-    #[test]
-    fn test_columnar_hash_join_left_outer() {
-        // Create left batch: customer_id, name (4 customers)
-        let left_columns = vec![
-            ColumnArray::Int64(Arc::new(vec![1, 2, 3, 4]), None),
-            ColumnArray::String(Arc::new(vec!["Alice".into(), "Bob".into(), "Carol".into(), "Dave".into()]), None),
-        ];
-        let left_batch = ColumnarBatch::from_columns(left_columns, Some(vec!["customer_id".into(), "name".into()])).unwrap();
-
-        // Create right batch: order_id, customer_id (only customers 1, 2, 3 have orders)
-        let right_columns = vec![
-            ColumnArray::Int64(Arc::new(vec![101, 102, 103]), None),
-            ColumnArray::Int64(Arc::new(vec![1, 2, 1]), None), // Dave (id=4) has no orders
-        ];
-        let right_batch = ColumnarBatch::from_columns(right_columns, Some(vec!["order_id".into(), "customer_id".into()])).unwrap();
-
-        // LEFT OUTER JOIN on customer_id (left col 0 = right col 1)
-        let result = columnar_hash_join_left_outer(&left_batch, &right_batch, 0, 1).unwrap();
-
-        // Should have 5 result rows:
-        // - Alice (1): 2 matches (101, 103)
-        // - Bob (2): 1 match (102)
-        // - Carol (3): 0 matches, but preserved with NULLs
-        // - Dave (4): 0 matches, but preserved with NULLs
-        assert_eq!(result.row_count(), 5);
-        assert_eq!(result.column_count(), 4); // 2 left + 2 right
-
-        // Verify that all left rows are preserved
-        let rows = result.to_rows().unwrap();
-
-        // Count customers preserved
-        let mut customer_counts = std::collections::HashMap::new();
-        for row in &rows {
-            if let Some(vibesql_types::SqlValue::Integer(id)) = row.get(0) {
-                *customer_counts.entry(*id).or_insert(0) += 1;
-            }
-        }
-
-        // Alice should appear 2 times, Bob 1 time, Carol 1 time, Dave 1 time
-        assert_eq!(customer_counts.get(&1), Some(&2)); // Alice
-        assert_eq!(customer_counts.get(&2), Some(&1)); // Bob
-        assert_eq!(customer_counts.get(&3), Some(&1)); // Carol (preserved with NULL)
-        assert_eq!(customer_counts.get(&4), Some(&1)); // Dave (preserved with NULL)
-    }
-
-    #[test]
-    fn test_columnar_hash_join_right_outer() {
-        // Create left batch: customer_id, name (2 customers)
-        let left_columns = vec![
-            ColumnArray::Int64(Arc::new(vec![1, 2]), None),
-            ColumnArray::String(Arc::new(vec!["Alice".into(), "Bob".into()]), None),
-        ];
-        let left_batch = ColumnarBatch::from_columns(left_columns, Some(vec!["customer_id".into(), "name".into()])).unwrap();
-
-        // Create right batch: order_id, customer_id (customer 3 has no matching left row)
-        let right_columns = vec![
-            ColumnArray::Int64(Arc::new(vec![101, 102, 103, 104]), None),
-            ColumnArray::Int64(Arc::new(vec![1, 2, 3, 1]), None), // Order 103 has customer_id=3, not in left
-        ];
-        let right_batch = ColumnarBatch::from_columns(right_columns, Some(vec!["order_id".into(), "customer_id".into()])).unwrap();
-
-        // RIGHT OUTER JOIN on customer_id (left col 0 = right col 1)
-        let result = columnar_hash_join_right_outer(&left_batch, &right_batch, 0, 1).unwrap();
-
-        // Should have 4 result rows (all right rows preserved):
-        // - Order 101 (customer 1): matches Alice
-        // - Order 102 (customer 2): matches Bob
-        // - Order 103 (customer 3): no match, left columns are NULL
-        // - Order 104 (customer 1): matches Alice
-        assert_eq!(result.row_count(), 4);
-        assert_eq!(result.column_count(), 4); // 2 left + 2 right
-
-        // Verify that all right rows are preserved
-        let rows = result.to_rows().unwrap();
-
-        // Count by order_id (column 2 in result)
-        let mut order_found = std::collections::HashSet::new();
-        let mut null_customer_count = 0;
-
-        for row in &rows {
-            if let Some(vibesql_types::SqlValue::Integer(order_id)) = row.get(2) {
-                order_found.insert(*order_id);
-            }
-            if let Some(vibesql_types::SqlValue::Null) = row.get(0) {
-                null_customer_count += 1;
-            }
-        }
-
-        // All 4 orders should be present
-        assert!(order_found.contains(&101));
-        assert!(order_found.contains(&102));
-        assert!(order_found.contains(&103));
-        assert!(order_found.contains(&104));
-
-        // One row should have NULL customer (order 103)
-        assert_eq!(null_customer_count, 1);
-    }
-
-    #[test]
-    fn test_columnar_hash_join_with_nulls() {
-        // Create left batch with NULL key
-        let left_columns = vec![
-            ColumnArray::Int64(
-                Arc::new(vec![1, 0, 3]), // 0 is placeholder for NULL
-                Some(Arc::new(vec![false, true, false])), // Index 1 is NULL
-            ),
-            ColumnArray::String(Arc::new(vec!["Alice".into(), "Bob".into(), "Carol".into()]), None),
-        ];
-        let left_batch = ColumnarBatch::from_columns(left_columns, Some(vec!["id".into(), "name".into()])).unwrap();
-
-        // Create right batch
-        let right_columns = vec![
-            ColumnArray::Int64(Arc::new(vec![1, 3]), None),
-            ColumnArray::String(Arc::new(vec!["Order1".into(), "Order3".into()]), None),
-        ];
-        let right_batch = ColumnarBatch::from_columns(right_columns, Some(vec!["id".into(), "desc".into()])).unwrap();
-
-        // LEFT OUTER JOIN - Bob (NULL key) should be preserved with NULL right columns
-        let result = columnar_hash_join_left_outer(&left_batch, &right_batch, 0, 0).unwrap();
-
-        // Should have 3 rows: Alice matches, Bob (NULL key) preserved, Carol matches
-        assert_eq!(result.row_count(), 3);
-
-        // Verify Bob's row has NULL for right side
-        let rows = result.to_rows().unwrap();
-        let bob_row = rows.iter().find(|r| {
-            matches!(r.get(1), Some(vibesql_types::SqlValue::Varchar(s)) if s == "Bob")
-        });
-
-        assert!(bob_row.is_some());
-        let bob = bob_row.unwrap();
-        // Bob's right-side columns should be NULL
-        assert!(matches!(bob.get(2), Some(vibesql_types::SqlValue::Null)));
-        assert!(matches!(bob.get(3), Some(vibesql_types::SqlValue::Null)));
-    }
-
-    #[test]
-    fn test_composite_int_hash_table_build_and_probe() {
-        // Build table with two columns
-        let col1 = vec![1i64, 2, 3, 1];
-        let col2 = vec![10i64, 20, 30, 10]; // Row 0 and 3 have same composite key (1, 10)
-        let columns: Vec<&[i64]> = vec![col1.as_slice(), col2.as_slice()];
-
-        let ht = CompositeIntHashTable::build_from_multi_i64(&columns);
-
-        // Probe for (1, 10) - should find indices 0 and 3
-        let probe_key = vec![1i64, 10];
-        let matches: Vec<u32> = ht.probe_multi_i64(&probe_key, &columns).collect();
-        assert_eq!(matches.len(), 2);
-        assert!(matches.contains(&0));
-        assert!(matches.contains(&3));
-
-        // Probe for (2, 20) - should find index 1
-        let probe_key = vec![2i64, 20];
-        let matches: Vec<u32> = ht.probe_multi_i64(&probe_key, &columns).collect();
-        assert_eq!(matches.len(), 1);
-        assert!(matches.contains(&1));
-
-        // Probe for (1, 20) - should find nothing (partial match)
-        let probe_key = vec![1i64, 20];
-        let matches: Vec<u32> = ht.probe_multi_i64(&probe_key, &columns).collect();
-        assert!(matches.is_empty());
-
-        // Probe for (99, 99) - should find nothing
-        let probe_key = vec![99i64, 99];
-        let matches: Vec<u32> = ht.probe_multi_i64(&probe_key, &columns).collect();
-        assert!(matches.is_empty());
-    }
-
-    #[test]
-    fn test_hash_join_indices_columnar_multi() {
-        use vibesql_storage::Row;
-        use vibesql_types::SqlValue;
-
-        // Build rows: (a, b) pairs
-        let build_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(10)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(20)]),
-            Row::new(vec![SqlValue::Integer(3), SqlValue::Integer(30)]),
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(10)]), // Duplicate key
-        ];
-
-        // Probe rows: (a, b) pairs
-        let probe_rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Integer(10)]), // Matches build[0] and build[3]
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Integer(20)]), // Matches build[1]
-            Row::new(vec![SqlValue::Integer(99), SqlValue::Integer(99)]), // No match
-        ];
-
-        let pairs = hash_join_indices_columnar_multi(
-            &build_rows, &probe_rows, &[0, 1], &[0, 1]
-        ).unwrap();
-
-        // Should have 3 pairs: (0,0), (3,0), (1,1)
-        assert_eq!(pairs.len(), 3);
-
-        // Probe row 0 matches build rows 0 and 3
-        let matches_for_probe_0: Vec<_> = pairs.iter().filter(|(_, p)| *p == 0).collect();
-        assert_eq!(matches_for_probe_0.len(), 2);
-
-        // Probe row 1 matches build row 1
-        let matches_for_probe_1: Vec<_> = pairs.iter().filter(|(_, p)| *p == 1).collect();
-        assert_eq!(matches_for_probe_1.len(), 1);
-        assert_eq!(matches_for_probe_1[0].0, 1);
-
-        // Probe row 2 has no matches
-        let matches_for_probe_2: Vec<_> = pairs.iter().filter(|(_, p)| *p == 2).collect();
-        assert_eq!(matches_for_probe_2.len(), 0);
-    }
-
-    #[test]
-    fn test_hash_join_indices_columnar_multi_with_non_integer() {
-        use vibesql_storage::Row;
-        use vibesql_types::SqlValue;
-
-        // Build rows with string column (should fall back to None)
-        let build_rows = vec![
-            Row::new(vec![SqlValue::Varchar("a".to_string()), SqlValue::Integer(10)]),
-        ];
-
-        let probe_rows = vec![
-            Row::new(vec![SqlValue::Varchar("a".to_string()), SqlValue::Integer(10)]),
-        ];
-
-        // Should return None because not all columns are integers
-        let result = hash_join_indices_columnar_multi(
-            &build_rows, &probe_rows, &[0, 1], &[0, 1]
-        );
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_extract_multi_i64_columns() {
-        use vibesql_storage::Row;
-        use vibesql_types::SqlValue;
-
-        let rows = vec![
-            Row::new(vec![SqlValue::Integer(1), SqlValue::Bigint(100), SqlValue::Smallint(10)]),
-            Row::new(vec![SqlValue::Integer(2), SqlValue::Bigint(200), SqlValue::Smallint(20)]),
-        ];
-
-        // Extract columns 0 and 2
-        let columns = extract_multi_i64_columns(&rows, &[0, 2]).unwrap();
-        assert_eq!(columns.len(), 2);
-        assert_eq!(columns[0], vec![1i64, 2]);
-        assert_eq!(columns[1], vec![10i64, 20]);
-
-        // Extract all columns
-        let columns = extract_multi_i64_columns(&rows, &[0, 1, 2]).unwrap();
-        assert_eq!(columns.len(), 3);
-        assert_eq!(columns[0], vec![1i64, 2]);
-        assert_eq!(columns[1], vec![100i64, 200]);
-        assert_eq!(columns[2], vec![10i64, 20]);
-    }
-}
